Fix circular dep
(cherry picked from commit 9e7da8d54075ba1b4bd0ee6cda438adf055e03af)
